requirements.txt
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
model-bakery==1.9.0
Pillow==9.3.0
psycopg2==2.9.5
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.6

